refactor: move MyAnnoucements view styles to separate file
