+++ src/views/MyAnnoucements/index.tsx
@@ -1,4 +1,5 @@
-import { Box, createStyles, makeStyles, Theme, Typography } from '@material-ui/core';
+import Box from '@material-ui/core/Box';
+import Typography from '@material-ui/core/Typography';
 import { useEffect, useState } from 'react';
 import Card from '../../components/ProductCard/card';
 import { Annoucements, annoucements as ann } from '../../data/annoucements';
@@ -7,28 +8,7 @@ import AddIcon from '@material-ui/icons/Add';
 import { Link } from 'react-router-dom';
 import ViewTitle from '../../components/ViewTitle';
 import ViewContainer from '../../components/ViewContainer';
-
-const useStyles = makeStyles((theme: Theme) =>
-  createStyles({
-    addItemBox: {
-      background: '#E5E7EB',
-      color: '#9CA3AF',
-      display: 'grid',
-      placeItems: 'center',
-      borderRadius: theme.shape.borderRadius,
-      '& div': {
-        display: 'flex',
-        flexDirection: 'column',
-        alignItems: 'center',
-      },
-    },
-    plusIcon: {
-      width: '25%',
-      height: '25%',
-      strokeWidth: '1px',
-    },
-  })
-);
+import useStyles from './styles';
 
 const MyAnnoucements = () => {
   const [annoucements, setAnnoucements] = useState<Annoucements[]>([]);
